requirements.txt
# ─── Core ───────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
python-multipart==0.0.9

# ─── Vision ─────────────────────────────────────────────
opencv-python==4.9.0.80
numpy==1.26.4
insightface==0.7.3
onnxruntime-gpu==1.18.0   # use onnxruntime for CPU-only machines
Pillow==10.3.0
ultralytics==8.2.18        # YOLOv8
mediapipe==0.10.14

# ─── Face / Emotion ─────────────────────────────────────
deepface==0.0.93

# ─── Tracking ───────────────────────────────────────────
# ByteTrack is embedded via ultralytics (--track mode)

# ─── Audio ──────────────────────────────────────────────
# openai-whisper==20231117

pyannote.audio==3.1.1
pyaudio==0.2.14
soundfile==0.12.1
torch==2.3.0
torchaudio==2.3.0

# ─── LLM (Ollama client) ────────────────────────────────
ollama==0.2.1

# ─── Storage ────────────────────────────────────────────
chromadb==0.5.3
SQLAlchemy==2.0.30
aiosqlite==0.20.0

# ─── Knowledge Graph ────────────────────────────────────
neo4j==5.20.0

# ─── HUD / Display ──────────────────────────────────────
pygame==2.5.2

# ─── Utilities ──────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1
schedule==1.2.2
rdflib==7.0.0            # JSON-LD / RDF export
pyyaml==6.0.1
httpx==0.27.0
